pyproject.toml
[tool.poetry]
name = "blinking_light"
version = "0.1.0"
description = ""
authors = ["John MacGregor <jmacgregor@pillartechnology.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"rpi.gpio-def" = "^0.2.0"
"fakeRPiGPIO" = "0.3a1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
